+++ Categories/Foundation/NSDate+LjsAdditions.m
@@ -180,7 +180,7 @@ NSSecondCalendarUnit);
   
   NSDateComponents *difference = [aCalendar components:NSDayCalendarUnit
                                               fromDate:fromDate toDate:toDate options:0];
-  return abs(difference.day);
+  return (NSUInteger)abs(difference.day);
 }
 
 - (NSDate *) dateByAddingDays:(NSInteger) aNumberOfDays {
